Module mausam::app::ipapi

Structs

IpApiResponse 🔒

__FieldVisitor 🔒

__Visitor 🔒

Enums

__Field 🔒

Constants

FIELDS 🔒

Functions

get_ip_api_location 🔒

get_ip_api_location fetches the current ip location.

get_location 🔒

get_location makes a GET request to the ip-api API and retrieves the location information in JSON format. The JSON response is then deserialized into a struct IpApiResponse using the serde crate.

try_ipapi_location 🔒

let location = try_ipapi_location(client, *IP_API_URL).await?; Note that most geolocation APIs have usage limits, so be mindful of how often you make requests to the API.

Type Definitions

ResultIpApi 🔒

Opaque Types

🔒

🔒

🔒